Added thesis_result.m, further data management improvements
Removed the need to save greens function entierly for results of simulation which further reduce required data size, also started generating concrete results to the thesis work. Finally the NEGF_param was introduced which is an object with all the parameters that is to be fed to the NEGF simulations to make code more elegant and easier to manipulate.